+++ controllers/meController.js
@@ -2,12 +2,14 @@ const SerMe = require('../services/serMe')
 
 const getMyInfo = async (req, res) => {
   const myService = new SerMe()
-  const myInfo = await myService.findMyInfo(req.token.id)
-  // const myInfo = await myService.findMyInfo(1);
+  //const myInfo = await myService.findMyInfo(req.token.id)
+  const myTeamList = await myService.findMyTeamList(2);
+  const myInfo = await myService.findMyInfo(2);
   res.json({
     status: 200,
     data: {
-      myInfo
+      myInfo,
+      myTeamList
     }
   })
 }
