+++ models/modTeam.js
@@ -20,17 +20,16 @@ class ModelTeam {
   // 인원에 따른 팀 리스트 찾기
 
   // 나의 개별 팀 리스트 찾기
-  async findMyTeamList (user_id) {
+  async findMyTeamList (userId) {
     const teamList = []
     await Team.findAll({
       attributes: ['id'],
       where: {
-        owner_id: user_id
+        owner_id: userId
       }
     }).then(teams => {
       teams.map(team => teamList.push(team.dataValues.id))
     })
-    console.log('mod:', teamList)
     return teamList
   }
