+++ package.json
@@ -18,6 +18,7 @@
         "express": "^4.17.1",
         "express-mysql-session": "^2.1.0",
         "express-session": "^1.17.0",
+        "ipaddr": "0.0.9",
         "jsonwebtoken": "^8.5.1",
         "logger": "0.0.1",
         "morgan": "^1.9.1",
